test(loading): migrate is-open
